requirements.txt
discord.py==2.1.0
feedparser==6.0.10
protobuf==4.21.12
python_dateutil==2.8.2
pytz==2022.7
